Events: make the association explicit
This is a big change in how we do things, but I think it's a good one. Basically, we want to no longer have a complex query for events, we just want a basic, explicit association from vulnerability to event. That pushes the computation time for the rebuild script, but it saves time and complexity up front.

I don't want the rebuild script to be massively slow, but I don't think there's a better way to do this.