vulnerability#show: eager-load details and styles
This means that these are included in the original join for getting vulnerablity events, instead of having to load them one-by-one from the beginning.

I *think* this leads to a performance improvement. Hard to see in the development environment, though. Certainly simplifies the logs since there are fewer calls now