performance: use low-level caching on all API calls
This change is about assuming that all API calls will return the same JSON on a GET request once the data is built.

On a data rebuild, we'll clear the cache. And the cache will also expire in 12 hours just in case - but we can tweak this.

Caching is disabled on dev, unless you do the rails dev:cache. In production we'll use whatever we can, probably Redis.