VHP-flavored markdown links to models (#427) (#493)
I've added support for icons. Right now, only CVEs are auto-linked and are displayed with a bug icon to the left of the link. Minor aesthetic changes will need to be made to how icons are displayed.

Depending on how we decide to implement the linking of the other models (tags, filepaths, etc.), we might be able to just add more regex expressions and other logic in the application helper. If we decide to not do any special keyword extraction from the text and just use Markdown and Kramdown as discussed on #427, then we may not be using any of this.