Hide email address on developer#show behind AJAX call (#509)
Since the developer page was previously unable to load, I had to take out some code relating to the event-detail associations before I could start working on issue #506.

With this pull request, a developer's email address will not be displayed on the page until either mouse movement or keyboard input is detected from the user, and the user clicks on the "Fetch Email" button. When the user clicks that button, an AJAX call is made to /api/developers/<id> which returns their email address, and then that address is injected into the page.