Improve display of table on small devices (#529) (#643)
I made the datatable responsive, like how it is on vulnerabilities#index. However, because a responsive datable adds a clickable plus button to a row when the screen size shrinks, I had to change how clicking is handled. Now, instead of the entire row being a link that'll take you to the commits page, each string of text is a link, and the plus button will expand to show you any hidden columns. Without this change, clicking the plus button would inadvertently take you to the commits page. (#529)

I also fixed some inconsistent spacing and tabbing.

* Improve display of table on small devices (#529)

* Responsively resize datatable (#529)