We have news! (#1022)
This closes #1018 I wanted some basic changelog functionality. Something that we can handwrite but also use in multiple places on the website.

It's pretty basic but a nice addition I think. We can work on UI in #1020 - this is mostly to get the functionality presentable.

Also commented out some articles links from the top bar - trying to de-emphasize that